+++ config/initializers/tag_ext.rb
@@ -1,3 +1,7 @@
+# Make tags space-delimited.
+TagList.delimiter = " "
+
+# Add a #to_param method to tags.
 module TagExt
 
   def to_param