+++ CHANGES.md
@@ -5,6 +5,7 @@ r0.0.7.6
 --------
 
 * Fixed module tags, they weren't being assigned.
+* Changed tags so that they're space-delimited, rather than comma delimited.
 * Disabled authentication via HTTP Basic Auth so that the site can run with .htaccess protection.
 
 r0.0.7.5